Greatly improved processing speed for backend calc by removing if checks
256Tile-ROI CoreBx (HER2, CK, DAPI)
CPU (32 threads): 10s
CUDA GPU (RTX4090): 7s

The boilerplate could be cleaned up, but is very fast!
Need to add option to enable/disable CUDA in the settings and test for CUDA enabled GPU at install.
Still need to look at the "Run for project" thread utilization, it does not seem to allocate the correct number of workers for a Quant task.